+++ test/integrations/component.test.ts
@@ -66,6 +66,7 @@ const INTEGRATIONS_WITH_UPDATED_TEST_STRUCTURE = [
   'klaviyo',
   'campaign_manager',
   'criteo_audience',
+  'customerio_audience',
   'branch',
   'userpilot',
   'loops',
@@ -78,7 +79,13 @@ const INTEGRATIONS_WITH_UPDATED_TEST_STRUCTURE = [
   'attentive_tag',
 ];
 
-const STREAMING_DEST_WITH_UPDATED_TEST_STRUCTURE = ['googlesheets'];
+const STREAMING_DEST_WITH_UPDATED_TEST_STRUCTURE = [
+  'googlesheets',
+  'kafka',
+  'kinesis',
+  'personalize',
+  'eventbridge',
+];
 
 beforeAll(async () => {
   initaliseReport();