+++ test/integrations/destinations/am/dataDelivery/data.ts
@@ -66,8 +66,12 @@ export const data = [
           output: {
             status: 200,
             message: '[amplitude Response Handler] - Request Processed Successfully',
-            destinationResponse:
-              '{"code":200,"server_upload_time":1639235302252,"payload_size_bytes":863,"events_ingested":1}',
+            destinationResponse: {
+              code: 200,
+              server_upload_time: 1639235302252,
+              payload_size_bytes: 863,
+              events_ingested: 1,
+            },
           },
         },
       },
@@ -547,7 +551,239 @@ export const data = [
   },
   {
     name: 'am',
-    description: 'Test 6: for 429 Rate Limit Handling',
+    description: 'Test 6: for 429 Rate Limit Handling (ThrottledUsers)',
+    feature: 'dataDelivery',
+    module: 'destination',
+    version: 'v0',
+    input: {
+      request: {
+        body: {
+          type: 'REST',
+          endpoint: 'https://api.amplitude.com/2/httpapi/rate-limited',
+          method: 'POST',
+          userId: 'test_user_123',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: {
+            FORM: {},
+            JSON: {
+              api_key: 'dummy-api-key',
+              events: [
+                {
+                  app_name: 'Rudder-Amplitude_Example',
+                  app_version: '1.0',
+                  time: 1619006730330,
+                  user_id: 'testrluser@email.com',
+                  user_properties: {
+                    city: 'San Francisco',
+                    country: 'US',
+                    email: 'testrluser@email.com',
+                  },
+                },
+              ],
+              options: {
+                min_id_length: 1,
+              },
+            },
+            JSON_ARRAY: {},
+            XML: {},
+          },
+          files: {},
+          params: {
+            destination: 'am',
+          },
+        },
+        method: 'POST',
+      },
+    },
+    output: {
+      response: {
+        status: 500, // Expected status is 500 (RetryableError)
+        body: {
+          output: {
+            status: 500,
+            message:
+              'Request Failed during amplitude response transformation: Too many requests for some devices and users - due to Request Limit exceeded, (Retryable)',
+            destinationResponse: {
+              headers: {
+                'access-control-allow-methods': 'GET, POST',
+                'access-control-allow-origin': '*',
+                'content-type': 'application/json',
+                'retry-after': '120',
+              },
+              response: {
+                code: 429,
+                error: 'Too many requests for some devices and users',
+                eps_threshold: 30,
+                throttled_users: {
+                  'testrluser@email.com': 32,
+                },
+                throttled_events: [3, 4, 7],
+              },
+              status: 429,
+            },
+            statTags: {
+              destType: 'AM',
+              errorCategory: 'network',
+              destinationId: 'Non-determininable',
+              workspaceId: 'Non-determininable',
+              errorType: 'retryable',
+              feature: 'dataDelivery',
+              implementation: 'native',
+              module: 'destination',
+            },
+          },
+        },
+      },
+    },
+    mockFns: (mockAdapter: MockAdapter) => {
+      mockAdapter
+        .onPost('https://api.amplitude.com/2/httpapi/rate-limited', {
+          asymmetricMatch: (actual) => {
+            // Simple check to match the request body
+            return actual.api_key === 'dummy-api-key';
+          },
+        })
+        .replyOnce(
+          429,
+          {
+            code: 429,
+            error: 'Too many requests for some devices and users',
+            eps_threshold: 30,
+            throttled_users: {
+              'testrluser@email.com': 32,
+            },
+            throttled_events: [3, 4, 7],
+          },
+          {
+            'access-control-allow-methods': 'GET, POST',
+            'access-control-allow-origin': '*',
+            'content-type': 'application/json',
+            'retry-after': '120',
+          },
+        );
+    },
+  },
+  {
+    name: 'am',
+    description: 'Test 7: for 429 Rate Limit Handling (ThrottledDevices)',
+    feature: 'dataDelivery',
+    module: 'destination',
+    version: 'v0',
+    input: {
+      request: {
+        body: {
+          type: 'REST',
+          endpoint: 'https://api.amplitude.com/2/httpapi/rate-limited',
+          method: 'POST',
+          userId: 'test_user_123',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: {
+            FORM: {},
+            JSON: {
+              api_key: 'dummy-api-key',
+              events: [
+                {
+                  app_name: 'Rudder-Amplitude_Example',
+                  app_version: '1.0',
+                  time: 1619006730330,
+                  user_id: 'testrluser@email.com',
+                  user_properties: {
+                    city: 'San Francisco',
+                    country: 'US',
+                    email: 'testrluser@email.com',
+                  },
+                },
+              ],
+              options: {
+                min_id_length: 1,
+              },
+            },
+            JSON_ARRAY: {},
+            XML: {},
+          },
+          files: {},
+          params: {
+            destination: 'am',
+          },
+        },
+        method: 'POST',
+      },
+    },
+    output: {
+      response: {
+        status: 500, // Expected status is 500 (RetryableError)
+        body: {
+          output: {
+            status: 500,
+            message:
+              'Request Failed during amplitude response transformation: Too many requests for some devices and users - due to Request Limit exceeded, (Retryable)',
+            destinationResponse: {
+              headers: {
+                'access-control-allow-methods': 'GET, POST',
+                'access-control-allow-origin': '*',
+                'content-type': 'application/json',
+                'retry-after': '120',
+              },
+              response: {
+                code: 429,
+                error: 'Too many requests for some devices and users',
+                eps_threshold: 30,
+                throttled_devices: {
+                  'HIJ3L821-F01A-2GY5-2C81-7F03X7DS291D': 31,
+                },
+                throttled_events: [3, 4, 7],
+              },
+              status: 429,
+            },
+            statTags: {
+              destType: 'AM',
+              errorCategory: 'network',
+              destinationId: 'Non-determininable',
+              workspaceId: 'Non-determininable',
+              errorType: 'retryable',
+              feature: 'dataDelivery',
+              implementation: 'native',
+              module: 'destination',
+            },
+          },
+        },
+      },
+    },
+    mockFns: (mockAdapter: MockAdapter) => {
+      mockAdapter
+        .onPost('https://api.amplitude.com/2/httpapi/rate-limited', {
+          asymmetricMatch: (actual) => {
+            // Simple check to match the request body
+            return actual.api_key === 'dummy-api-key';
+          },
+        })
+        .replyOnce(
+          429,
+          {
+            code: 429,
+            error: 'Too many requests for some devices and users',
+            eps_threshold: 30,
+            throttled_devices: {
+              'HIJ3L821-F01A-2GY5-2C81-7F03X7DS291D': 31,
+            },
+            throttled_events: [3, 4, 7],
+          },
+          {
+            'access-control-allow-methods': 'GET, POST',
+            'access-control-allow-origin': '*',
+            'content-type': 'application/json',
+            'retry-after': '120',
+          },
+        );
+    },
+  },
+  {
+    name: 'am',
+    description: 'Test 8: for 429 Rate Limit Handling (ThrottledUsers and ThrottledDevices)',
     feature: 'dataDelivery',
     module: 'destination',
     version: 'v0',
@@ -620,7 +856,7 @@ export const data = [
                 },
                 throttled_events: [3, 4, 7],
               },
-              status: 200,
+              status: 429,
             },
             statTags: {
               destType: 'AM',
@@ -645,7 +881,7 @@ export const data = [
           },
         })
         .replyOnce(
-          200,
+          429,
           {
             code: 429,
             error: 'Too many requests for some devices and users',
@@ -669,7 +905,7 @@ export const data = [
   },
   {
     name: 'am',
-    description: 'Test 7: for standard 429 Rate Limit Handling (ThrottledError)',
+    description: 'Test 9: for standard 429 Rate Limit Handling (ThrottledError)',
     feature: 'dataDelivery',
     module: 'destination',
     version: 'v0',
@@ -734,7 +970,7 @@ export const data = [
                 throttled_events: [],
                 throttled_users: {},
               },
-              status: 200,
+              status: 429,
             },
             message:
               'Request Failed during amplitude response transformation: Rate limit exceeded - due to Request Limit exceeded, (Throttled)',
@@ -762,7 +998,7 @@ export const data = [
           },
         })
         .replyOnce(
-          200,
+          429,
           {
             code: 429,
             error: 'Rate limit exceeded',